prepare 4.1.0

+++ vertigo-javalin-connector/src/main/java/io/vertigo/connectors/javalin/JettyMultipartCleaner.java
@@ -22,6 +22,8 @@ import java.util.Collection;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
+import org.eclipse.jetty.http.HttpField;
+import org.eclipse.jetty.http.MimeTypes;
 
 import io.javalin.http.Context;
 import io.javalin.http.Handler;
@@ -38,17 +40,19 @@ public final class JettyMultipartCleaner implements Handler {
 	/** {@inheritDoc} */
 	@Override
 	public void handle(final Context ctx) {
-		Collection<Part> multiParts;
 		try {
-			multiParts = ctx.req().getParts();
-			if (multiParts != null && !multiParts.isEmpty()) {
-				for (final Part part : multiParts) {
-					try {
-						// a multipart request to a servlet will have the parts cleaned up correctly, but
-						// the repeated call to deleteParts() here will safely do nothing.
-						part.delete();
-					} catch (final IOException e) {
-						LOG.warn("Error while deleting multipart request parts", e);
+			final String contentType = ctx.req().getContentType();
+			if (contentType != null && MimeTypes.Type.MULTIPART_FORM_DATA.is(HttpField.valueParameters(contentType, null))) {
+				final Collection<Part> multiParts = ctx.req().getParts();
+				if (multiParts != null && !multiParts.isEmpty()) {
+					for (final Part part : multiParts) {
+						try {
+							// a multipart request to a servlet will have the parts cleaned up correctly, but
+							// the repeated call to deleteParts() here will safely do nothing.
+							part.delete();
+						} catch (final IOException e) {
+							LOG.warn("Error while deleting multipart request parts", e);
+						}
 					}
 				}
 			}
